fix: 修复变量类型

+++ sky-take-out/sky-server/src/main/java/com/sky/controller/admin/DishController.java
@@ -44,7 +44,7 @@ public class DishController {
      */
     @GetMapping("/list")
     @ApiOperation("根据分类id查询菜品")
-    public Result<List<DishVO>> pageList(@RequestParam Integer categoryId){
+    public Result<List<DishVO>> pageList(@RequestParam Long categoryId){
         log.info("根据分类id查询菜品");
         return dishService.listDish(categoryId);
     }
